Revert "feat(upgrade): Add terminal tab to client sandboxes for upgrade lever (#7394)"
This reverts commit baaa7585dbf230a03bda6cdfea40507e167e4841.
